Have trace() and debug() really eliminated with NDEBUG
The OutputFormatter::operator<< wasn't eliminated, as it is too large to
get inlined. Thus, the check for out != 0 happens in the function and it
must be called with all the stream operator arguments.
By using a DummyOutputFormatter object with an empty stream operator the
compiler can really eliminate the calls.

While at it I simplified the be*() functions, which contained a bit of
code duplication.

Signed-off-by: Matthias Kretz <kretz@kde.org>