new log printing overload that can be optimized better
In addition to printing via stream operators, all print outputs now
support passing the arguments to be printed as function arguments to a
single function call. This way the enabled check has to be one only once
for every argument. Also the calling code becomes more concise.

before:
  log.debug() << foo << bar;
now (in addition to the above):
  log.debug(foo, bar);

Signed-off-by: Matthias Kretz <kretz@kde.org>